+++ src/js/Helper/Import/CsvConversionHelper.js
@@ -415,6 +415,11 @@ export default class ImportCsvConversionHelper {
                 firstLine: 1,
                 db       : 'passwords',
                 mapping  : ['label', 'url', '', 'username', 'password', 'notes']
+            },
+            safeincloud : {
+                firstLine: 1,
+                db       : 'passwords',
+                mapping  : ['label', 'username', 'password', 'url', 'notes']
             }
         };