+++ src/vue/Components/Import.vue
@@ -17,7 +17,8 @@
                     <translate tag="option" value="lastpass" say="LastPass CSV"/>
                     <translate tag="option" value="dashlane" say="Dashlane CSV"/>
                     <translate tag="option" value="roboform" say="RoboForm CSV"/>
-                    <translate tag="option" value="csv" say="Custom CSV"/>
+                    <translate tag="option" value="safeincloud" say="SafeInCloud CSV"/>
+                    <translate tag="option" value="csv" say="Other / Custom CSV"/>
                 </select>
             </div>
         </div>
@@ -377,6 +378,7 @@
                     case 'legacy':
                     case 'roboform':
                     case 'lastpass':
+                    case 'safeincloud':
                         this.options.profile = value;
                         break;
                     case 'pmanCsv':