+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
  - Sharing now available from menu
  - Support user hiding in guests app
+ - SafeInCloud CSV import
 ### Changed
  - Login screen now skipped entirely if not needed